Refactor to feature 'idiomatic' first, with full back-compat with 'astrapy' (#251)
Restructured package to feature 'idiomatic' first and foremost, keeping full back-compat with 'astrapy'
