+++ .github/workflows/main.yml
@@ -35,4 +35,4 @@ jobs:
 
     - name: Run pytest
       run: |
-        poetry run pytest tests/astrapy
+        poetry run pytest tests/core
