+++ tests/core/test_async_db_ddl.py
@@ -22,8 +22,8 @@ import pytest
 
 from ..conftest import AstraDBCredentials
 from .conftest import TEST_SKIP_COLLECTION_DELETE
-from astrapy.db import AsyncAstraDB, AsyncAstraDBCollection
-from astrapy.defaults import DEFAULT_KEYSPACE_NAME
+from astrapy.core.db import AsyncAstraDB, AsyncAstraDBCollection
+from astrapy.core.defaults import DEFAULT_KEYSPACE_NAME
 
 TEST_CREATE_DELETE_VECTOR_COLLECTION_NAME = "ephemeral_v_col"
 TEST_CREATE_DELETE_NONVECTOR_COLLECTION_NAME = "ephemeral_non_v_col"
