+++ tests/core/test_conversions.py
@@ -19,8 +19,13 @@ Tests for the User-Agent customization logic
 import logging
 import pytest
 
-from astrapy.db import AstraDB, AstraDBCollection, AsyncAstraDB, AsyncAstraDBCollection
-from astrapy.ops import AstraDBOps
+from astrapy.core.db import (
+    AstraDB,
+    AstraDBCollection,
+    AsyncAstraDB,
+    AsyncAstraDBCollection,
+)
+from astrapy.core.ops import AstraDBOps
 
 
 logger = logging.getLogger(__name__)
