+++ tests/core/test_db_ddl.py
@@ -22,8 +22,8 @@ import pytest
 
 from ..conftest import AstraDBCredentials
 from .conftest import TEST_SKIP_COLLECTION_DELETE
-from astrapy.db import AstraDB, AstraDBCollection
-from astrapy.defaults import DEFAULT_KEYSPACE_NAME
+from astrapy.core.db import AstraDB, AstraDBCollection
+from astrapy.core.defaults import DEFAULT_KEYSPACE_NAME
 
 TEST_CREATE_DELETE_VECTOR_COLLECTION_NAME = "ephemeral_v_col"
 TEST_CREATE_DELETE_NONVECTOR_COLLECTION_NAME = "ephemeral_non_v_col"
