+++ tests/core/test_imports.py
@@ -1,0 +1,31 @@
+# Copyright DataStax, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""
+Tests for the User-Agent customization logic
+"""
+
+import pytest
+
+
+@pytest.mark.describe("core imports")
+def test_core_imports() -> None:
+    from astrapy.api import APIRequestError  # noqa: F401
+    from astrapy.ops import AstraDBOps  # noqa: F401
+    from astrapy.db import (  # noqa: F401
+        AstraDBCollection,
+        AsyncAstraDBCollection,
+        AstraDB,
+        AsyncAstraDB,
+    )
